[Headline] Fixed image caption alignment
no refs.

Image caption was aligned to the right which looks cool on wide images but is very odd when the image doesn't go all the way to the edges of the content area — e.g. with gif's. The simple and quick solution is to align image captions to the center.